+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - QISKit parser does not halt at dictionary data ([#135](https://github.ibm.com/IBMQuantum/qiskit-studio/pull/135) by [@Yeray-Darias](https://github.ibm.com/Yeray-Darias))
 
 ### ✏️ Changed
+- No magic numbers at the QISKit symbol table ([#148](https://github.ibm.com/IBMQuantum/qiskit-studio/pull/148) by [@Yeray-Darias](https://github.ibm.com/Yeray-Darias))
 
 ### 👾 Security